Client for ARI.
EventEmitter Client that emits events for ARI web socket messages and
provides a wrapper for ARI resources and resource operations.

Events and API calls receive first class objects representing API
resources which can be used to call further API operations.

When using an instance of a resource, API calls automatically pass
the instance's identifier(s).

Instances can be created by the client, which generates a unique ID
for the purposes of setting up event listeners before the instance
is created in ARI. These event listerners are scoped to the
instance.