Connecting to unreachable host now returns and error.
Fixed a bug where attempting to connect to an unreachable host
would not indicate that an error occured.

This is due to the fact that the swagger client reports errors
once connected (authentication) but nothing at a lower level.
Instead we first connect directly to see if the host is
reachable, and then use swagger to connect if successful.