Add support for the 'Asterisk' resource
This patch adds support for the 'Asterisk' resource. This includes the
existing Asterisk resource operations, as well as the proposed push
configuration operations.

Note that since the 'Asterisk' resource does not have an identifier and
will always have a single 'instance' of itself, it does not need an ID,
nor does it really make use of the _params field.