lib/client.js: Add an option to the 'start' API call to subscribe to all events
A proposed patch to Asterisk (which would be released in 13.6.0) adds the
ability for a WebSocket connection to suscribe to all event sources in Asterisk
on connection. This patch adds support for that by adding a new option to the
'start' method. If true, the client will pass this along in its initial URI
used to establish the WebSocket, resulting in the application being subscribed
to all event sources.