Client can now manage event listeners for some resources.
For channels, bridges, live recordings, and playbacks, the client is now
able to manage the life cycle of event listeners if an instance is
explicitly marked for management using instance.manageInstance(). For
such instances, any event listener registered against the instance will
be unregistered once a cleanup event is received by the client.