Update dev dependencies
This patch updates all outdated development dependencies, but does
not upgrade the two outdated production dependencies.

- swagger-client, which is currently pinned at 2.1.17 due to a
breaking change that is difficult to address (see #47 for details).

- bluebird, which currently has a major version upgrade available
that includes breaking changes. See
http://bluebirdjs.com/docs/new-in-bluebird-3.html for details.
These are workable changes for node-ari-client, but we expose
bluebird Promises on our public API, so updating this would require
a major version bump.

For changes in the development dependencies, see the following
changelog entries:

- grunt: http://gruntjs.com/blog/2016-04-04-grunt-1.0.0-released
- async: https://github.com/caolan/async/blob/master/CHANGELOG.md
- mocha: https://github.com/mochajs/mocha/blob/master/CHANGELOG.md
- mustache: https://github.com/janl/mustache.js/blob/master/CHANGELOG.md
- portfinder: https://github.com/indexzero/node-portfinder/pull/20

No details regarding the major version bump of hock could be found,
but it had a substantial change to its public api. Namely, it does not
handle any of the actual http server stuff for you anymore. Changes
are included in this PR to adapt the tests to this new api.

There are still some npm install warnings even after updating these
dependencies:

- jshint is throwing a warning about minimatch needing an upgrade.
The fix for this is already in the jshint master branch, but no
npm version has been cut for it yet. For additional details, see
https://github.com/jshint/jshint/issues/2952

- jsdoc is throwing warnings about minimatch needing an upgrade,
wrench being completely deprecated in favor of another library,
and the marked package not being meant to be installed as a dep.
Upon investigation, it seems that the jsdoc package has actually
been abandoned, so we need to start thinking about migrating away
from it. Ticket #55 has been opened to investigate this.

- grunt has a warning about coffee-script not being meant to be
installed as a dep. This isn't really worth looking into, as we
are gradually moving away from grunt anyway.